requirements.txt
schedule==1.1.0
GitPython==3.1.30
python-dotenv==0.21.0
